fix: split graphql-codegen into types.ts and operations.ts to eliminate duplicate input types
- types.ts: schema types only (typescript plugin)
- operations.ts: operation types + documents (typescript-operations + typed-document-node)
- Updated imports in 6 files to use new @/app/graphql/operations path
- Added ObjectId conversion in companion-request.service.ts
