+++ package.json
@@ -17,6 +17,7 @@
     },
     "devDependencies": {
         "@graphql-codegen/cli": "^7.1.2",
+        "@graphql-codegen/import-types-preset": "^4.0.1",
         "@graphql-codegen/typed-document-node": "^7.0.3",
         "@graphql-codegen/typescript": "^6.0.2",
         "@graphql-codegen/typescript-operations": "^6.0.3",
